refactor: Update refreshToken options
