+++ pic-domain/src/main/kotlin/com/mashup/pic/domain/auth/RefreshTokenRepository.kt
@@ -3,7 +3,7 @@ package com.mashup.pic.domain.auth
 import org.springframework.data.repository.CrudRepository
 
 
-interface RefreshTokenRepository : CrudRepository<RefreshToken?, String?> {
+interface RefreshTokenRepository : CrudRepository<RefreshToken, String> {
     fun findByRefreshToken(refreshToken: String): RefreshToken?
-    fun deleteByRefreshToken(refreshToken: String?)
+    fun deleteByRefreshToken(refreshToken: String)
 }
