+++ pic-domain/src/main/kotlin/com/mashup/pic/domain/auth/RefreshTokenService.kt
@@ -3,12 +3,15 @@ package com.mashup.pic.domain.auth
 import com.mashup.pic.common.exception.PicException
 import com.mashup.pic.common.exception.PicExceptionType
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 
 @Service
+@Transactional(readOnly = true)
 class RefreshTokenService(
    private val refreshTokenRepository: RefreshTokenRepository
 ) {
 
+    @Transactional
     fun saveToken(userId: Long, token: String) {
         refreshTokenRepository.save(RefreshToken(token, userId));
     }
@@ -20,13 +23,10 @@ class RefreshTokenService(
         return refreshToken.userId
     }
 
+    @Transactional
     fun updateToken(userId: Long, originToken: String, newToken: String) {
         refreshTokenRepository.deleteByRefreshToken(originToken)
         refreshTokenRepository.save(RefreshToken(newToken, userId))
-    }
-
-    fun deleteToken(token: String) {
-        refreshTokenRepository.deleteByRefreshToken(token)
     }
 
 }
