+++ pic-domain/src/main/kotlin/com/mashup/pic/domain/config/RedisConfig.kt
@@ -1,12 +1,12 @@
-package com.mashup.pic.domain.common
+package com.mashup.pic.domain.config
 
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.cache.CacheManager
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.redis.cache.RedisCacheConfiguration
 import org.springframework.data.redis.cache.RedisCacheManager
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory
-import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer
 import org.springframework.data.redis.serializer.RedisSerializationContext
 import org.springframework.data.redis.serializer.StringRedisSerializer
@@ -14,24 +14,17 @@ import java.time.Duration
 
 
 @Configuration
-class RedisConfig {
-
-    @Bean
-    fun redisTemplate(): RedisTemplate<*, *> {
-        val redisTemplate = RedisTemplate<Any, Any>()
-        redisTemplate.setConnectionFactory(lettuceConnectionFactory())
-        redisTemplate.keySerializer = StringRedisSerializer()
-        redisTemplate.hashKeySerializer = StringRedisSerializer()
-        redisTemplate.valueSerializer = GenericJackson2JsonRedisSerializer()
-        return redisTemplate
-    }
+class RedisConfig(
+    @Value("\${redis.data.host}") private val host: String,
+    @Value("\${redis.data.port}") private val port: Int,
+) {
 
     @Bean
     fun cacheManager(): CacheManager {
         val configuration = RedisCacheConfiguration.defaultCacheConfig()
             .serializeKeysWith(RedisSerializationContext.SerializationPair.fromSerializer(StringRedisSerializer()))
             .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(GenericJackson2JsonRedisSerializer()))
-            .entryTtl(Duration.ofMinutes(2))
+            .entryTtl(Duration.ofDays(DEFAULT_ENTRY_TTL_DAYS))
             .disableCachingNullValues()
         return RedisCacheManager.RedisCacheManagerBuilder
             .fromConnectionFactory(lettuceConnectionFactory())
@@ -41,6 +34,10 @@ class RedisConfig {
 
     @Bean
     fun lettuceConnectionFactory(): LettuceConnectionFactory {
-        return LettuceConnectionFactory()
+        return LettuceConnectionFactory(host, port)
+    }
+
+    companion object {
+        private const val DEFAULT_ENTRY_TTL_DAYS: Long = 2
     }
 }
