+++ pic-domain/src/main/kotlin/com/mashup/pic/domain/user/UserRepository.kt
@@ -6,6 +6,4 @@ import org.springframework.stereotype.Repository
 @Repository
 interface UserRepository : JpaRepository<User, Long> {
     fun findByOAuthId(oauthId: Long): User?
-
-    fun findByIdOrNull(id: Long): User?
 }
