+++ pic-domain/src/main/kotlin/com/mashup/pic/domain/user/UserService.kt
@@ -2,6 +2,7 @@ package com.mashup.pic.domain.user
 
 import com.mashup.pic.common.exception.PicException
 import com.mashup.pic.common.exception.PicExceptionType
+import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -15,7 +16,7 @@ class UserService(
     }
 
     fun findUserByUserId(userId: Long): User {
-        return userRepository.findByIdOrNull(userId)?: throw PicException.of(PicExceptionType.AUTH_ERROR)
+        return userRepository.findByIdOrNull(userId) ?: throw PicException.of(PicExceptionType.AUTH_ERROR)
     }
 
     @Transactional
